Updated mav process
