+++ Interoperability/src/config.json
@@ -2,8 +2,8 @@
     "auvsi":{
         "host":"http://0.0.0.0",
         "port":"8000",
-	"username":"cnu_uas",
-	"password":"develop"
+	"username":"testadmin",
+	"password":"testpass"
     },
 
     "mavl-incoming":{
