+++ Interoperability/src/utils.py
@@ -17,6 +17,31 @@ class bcolors:
     	UNDERLINE = '\033[4m'
     	ERROR = '\033[1;41m'
 
+        
+#==================================
+#  Queue to handel the
+#  continuous posting process.
+#==================================
+class Queue():
+    "A container with a first-in-first-out (FIFO) queuing policy."
+    def __init__(self):
+        self.list = []
+
+    def push(self,item):
+        "Enqueue the 'item' into the queue"
+        self.list.insert(0,item)
+
+    def pop(self):
+        """
+          Dequeue the earliest enqueued item still in the queue. This
+          operation removes the item from the queue.
+        """
+        return self.list.pop()
+
+    def isEmpty(self):
+        "Returns true if the queue is empty"
+        return len(self.list) == 0
+
 
 #=============================
 #
@@ -34,6 +59,16 @@ class Utils():
 		print bcolors.BOLD + string + bcolors.ENDC
 		self.currLog = string
 		self.logs.append(string)
+        
+	def prompt(self, question, affirmOpt):
+		'''
+			Not functional yet. Will be used for CLI.
+		'''
+		opt = input(bcolors.BOLD + question + bcolors.ENDC)
+		if (opt in affirmOpt):
+			return True
+		else:
+			return False
 
 	def errLog(self, string):
 		print bcolors.ERROR + string + bcolors.ENDC
